Convert create-slot to typescript
